backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
openai>=1.0.0
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
pymavlink==2.4.37
pytest==7.4.3
httpx==0.25.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
google-generativeai>=0.3.0 